Refactor diplomacy planner into planning modules
Continue issue #2289 by moving diplomacy planning logic under src/planning and extracting war desire scoring helpers into a dedicated module while preserving public API compatibility through re-exports.
